+++ app/models/html_validation_scan.rb
@@ -2,6 +2,14 @@ require 'bashtetikn'
 
 # Scans a URL, recording HTML validation issues
 class HtmlValidationScan < PageScan
+  def self.icon_name
+    'microscope'
+  end
+
+  def icon_name
+    self.call.icon_name
+  end
+
   def scan
     results = validator.validate_uri(self.url)