+++ spec/routing/websites_routing_spec.rb
@@ -34,5 +34,9 @@ RSpec.describe WebsitesController, type: :routing do
     it "routes to #destroy" do
       expect(delete: "/websites/1").to route_to("websites#destroy", id: "1")
     end
+
+    it "routes to #html_validation_scan" do
+      expect(get: "/websites/1/html_validation_scan").to route_to("websites#html_validation_scan", id: "1")
+    end
   end
 end