+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 In place of release version numbers, we organize via deploys to Production (by Date/Time).
 
+## Upcoming: Scan all pages
+
+- install FontAwesome (https://github.com/tomkra/font_awesome5_rails)
+- add FontAwesomeHelper#font_awesome_icon
+- add PageScansHelper#page_scan_link_to
+- add rspec-html-matchers for rspec
+- use microscope icon for HtmlValidationScan
 
 ## 2021-11-19 HtmlValidationScans are async